requirements.txt
streamlit==1.43.2
ultralytics==8.4.25
opencv-python-headless==4.11.0.86
Pillow==11.1.0
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0